+++ bobashare-web/src/main.rs
@@ -76,7 +76,8 @@ async fn main() -> anyhow::Result<()> {
         .set_default("default_expiry", "24h").unwrap()
         .set_default("max_expiry", Some("30d")).unwrap()
         .set_default("max_file_size", 1024 * 1024 * 1024).unwrap() // 1 GiB
-        .set_default("extra_footer_text", None::<String>).unwrap();
+        .set_default("extra_footer_text", None::<String>).unwrap()
+        .set_default("about_page", None::<String>).unwrap();
 
     if let Some(c) = cli.config {
         config = config.add_source(config::File::new(
@@ -137,6 +138,17 @@ async fn main() -> anyhow::Result<()> {
     .map(|d| TimeDelta::from_std(d).unwrap());
     let max_file_size = config.get_int("max_file_size").unwrap().try_into().unwrap();
     let extra_footer_text = config.get("extra_footer_text").unwrap();
+    let about_page = config
+        .get::<Option<String>>("about_page")
+        .unwrap()
+        .map(|s| PathBuf::from(s));
+    let about_page_content = if let Some(ref path) = about_page {
+        // no reason to use tokio here since there is nothing to run concurrently yet
+        std::fs::read_to_string(path)
+            .with_context(|| format!("error reading about page contents at path {path:?}"))?
+    } else {
+        String::new()
+    };
 
     let state = Arc::new(AppState {
         backend,
@@ -151,6 +163,8 @@ async fn main() -> anyhow::Result<()> {
         syntax_set: SyntaxSet::load_defaults_newlines(),
 
         extra_footer_text,
+        about_page,
+        about_page_content,
 
         shutdown_tx: broadcast::channel(4).0,
     });
@@ -164,6 +178,7 @@ async fn main() -> anyhow::Result<()> {
         max_expiry = %state.max_expiry.map_or_else(|| String::from("never"), |e| e.to_string()),
         max_file_size = %state.max_file_size,
         extra_footer_text = ?state.extra_footer_text,
+        about_page = ?state.about_page,
         "generated state from config"
     );